split tests (#264)
* Split tests into multiple files

* rearrange tests

* reorganize tests

* move tests to flutter/tests app

* test fixes

* refactor tests

fix isReadOnly

* multiple jobs have a bug and don't report the names of the tests correctly. they report the same test name for all tests. That's only a bug in "-r expanded"

* hide test
