+++ .github/workflows/realm-dart-linux.yml
@@ -36,7 +36,7 @@ jobs:
 
       - name: Run tests
         run: |
-          dart test -r expanded --test-randomize-ordering-seed random
+          dart test -r expanded -j 1 --test-randomize-ordering-seed random
           echo "ARCHIVE_PATH=$(pwd)/binary/linux.tar.gz" >> $GITHUB_ENV
   
       - name: Archive binary
