+++ .github/workflows/realm-dart-macos.yml
@@ -37,7 +37,7 @@ jobs:
 
       - name: Run tests
         run: |
-          dart test -r expanded --test-randomize-ordering-seed random
+          dart test -r expanded -j 1 --test-randomize-ordering-seed random
           echo "ARCHIVE_PATH=$(pwd)/binary/macos.tar.gz" >> $GITHUB_ENV
 
       - name: Archive binary
