--- flutter/realm_flutter/test
@@ -1,1 +1,0 @@
-../../test
