+++ flutter/realm_flutter/tests/lib
@@ -1,0 +1,1 @@
+../lib
