+++ flutter/realm_flutter/tests/test
@@ -1,0 +1,1 @@
+../../../test
