+++ flutter/realm_flutter/tests/test_driver/realm_test.dart
@@ -6,13 +6,19 @@ import 'package:test_api/src/backend/invoker.dart';
 // ignore: implementation_imports
 import 'package:test_api/src/backend/state.dart' as test_api;
 
-import '../../test/realm_test.dart' as realm_tests;
+import '../test/configuration_test.dart' as configuration_test;
+import '../test/realm_test.dart' as realm_tests;
+import '../test/list_test.dart' as list_tests;
+import '../test/results_test.dart' as results_tests;
 
 Future<String> main(List<String> args) async {
   final Completer<String> completer = Completer<String>();
   final List<String> failedTests = [];
 
+  await configuration_test.main(args);
   await realm_tests.main(args);
+  await list_tests.main(args);
+  await results_tests.main(args);
 
   tearDown(() {
     if (Invoker.current?.liveTest.state.result == test_api.Result.error || Invoker.current?.liveTest.state.result == test_api.Result.failure) {
@@ -23,8 +29,7 @@ Future<String> main(List<String> args) async {
   tearDownAll(() {
     if (failedTests.isNotEmpty) {
       completer.complete(failedTests.join('\n'));
-    }
-    else {
+    } else {
       completer.complete('');
     }
   });
